# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Situation X - AI + Astrology Situation Analysis API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /analysis/estimate:
    post:
      operationId: estimateAnalysisCost
      tags: [analysis]
      summary: Exact credit cost of analysing this situation (modules involved + reasoning level), before anything is charged
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyzeRequest"
      responses:
        "200":
          description: Cost breakdown
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CostEstimate"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /analysis/analyze:
    post:
      operationId: analyzeSituation
      tags: [analysis]
      summary: Analyze a situation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyzeRequest"
      responses:
        "200":
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "402":
          description: Not enough credits (the body carries the price and the balance)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InsufficientCreditsResponse"
        "422":
          description: The ethical filter refused the request (it asks for something that could seriously harm someone). Nothing is analysed or charged; the message is a safe reply for the person.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /analysis/history:
    get:
      operationId: getAnalysisHistory
      tags: [analysis]
      summary: Get analysis history
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: History list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisHistoryResponse"
  /analysis/history/{id}/translate:
    post:
      operationId: translateAnalysis
      tags: [analysis]
      summary: Write the text of a finished analysis in another language (cached per language)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TranslateRequest"
      responses:
        "200":
          description: The translated texts, in the same structure as the analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranslateResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Translation is not available right now
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /analysis/history/{id}:
    get:
      operationId: getAnalysisById
      tags: [analysis]
      summary: Get analysis by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Analysis detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisRecord"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteAnalysis
      tags: [analysis]
      summary: Delete an analysis
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    AnalyzeRequest:
      type: object
      properties:
        situation:
          type: string
          description: The situation to analyze (min 10 characters)
        latitude:
          type: number
          description: Optional latitude of where the question is asked (default New Delhi). Only refines the Prashna ascendant; no birth details are needed.
        longitude:
          type: number
          description: Optional longitude of where the question is asked (default New Delhi)
        depth:
          type: string
          enum: [auto, standard, deep, expert]
          description: How deeply the AI should reason. "auto" (default) picks the level from how complex the question is. Deeper levels cost more credits.
        language:
          type: string
          enum: [auto, en, hi, hinglish]
          description: Language of the written answer. "auto" (default) follows the language of the question.
        useAi:
          type: boolean
          description: Use the AI for the final answer (default true). When false the engine and astrology answer is returned and no AI credits are charged.
        useAstrology:
          type: boolean
          description: Use the astrology lens (default true). When false the astrology is not used in the answer and is not charged.
        fromPath:
          type: object
          description: Run one of the paths SIVI listed in an earlier analysis of yours as a new question. The server builds the question from that analysis (the situation text is ignored) and charges half price. Looking at a path costs nothing.
          properties:
            analysisId:
              type: integer
            pathIndex:
              type: integer
              minimum: 0
              description: 0 is the recommended path, then the alternatives in order
          required:
            - analysisId
            - pathIndex
      required:
        - situation
    TranslateRequest:
      type: object
      properties:
        language:
          type: string
          enum: [en, hi, hinglish]
      required:
        - language
    TranslateResponse:
      type: object
      properties:
        language:
          type: string
          enum: [en, hi, hinglish]
        texts:
          type: object
          additionalProperties: true
          description: Translated sentences (summary, finalVerdict, synthesis, simulation, modules, astro)
      required:
        - language
        - texts
    ModuleReport:
      type: object
      description: What one module did for this question
      properties:
        key:
          type: string
          enum: [AJIT, MANU, FILTER, SIVI, ASTRO, AI]
        name:
          type: string
        area:
          type: string
          enum: [intent, emotion, paths, astrology, synthesis]
        role:
          type: string
          description: What the module looks at and how it understands the question
        active:
          type: boolean
        verdict:
          type: string
        evidence:
          type: array
          items:
            type: string
      required:
        - key
        - name
        - area
        - active
        - verdict
        - evidence
...
    IntentResult:
      type: object
      properties:
        intent:
          type: string
          enum: [decision, relationship, conflict, career, health, unclear]
        confidence:
          type: string
          enum: [low, medium, high]
        score:
          type: number
      required:
        - intent
        - confidence
        - score
    EmotionResult:
      type: object
      properties:
        emotion:
          type: string
          enum: [calm, stressed, anxious, angry, sad, confused]
        intensity:
          type: string
          enum: [low, medium, high]
        score:
          type: number
      required:
        - emotion
        - intensity
        - score
...
    SafetyNotice:
      type: object
      description: What the ethical filter decided about the message. "support" and "redirect" analyses carry a message that is shown to the person; blocked requests are refused with HTTP 422 and are never analysed or charged.
      properties:
        action:
          type: string
          enum: [allow, support, redirect, block]
        category:
          type: string
          enum: [none, self_harm, self_harm_method, harm_others, weapons, minor_sexual, manipulation]
        message:
          type: string
      required:
        - action
        - category
    SimulationResult:
      type: object
      description: SIVI comparison of the paths open to the person. It compares options; it does not predict the future.
      properties:
        context:
          $ref: "#/components/schemas/SituationContext"
        bestPath:
          $ref: "#/components/schemas/PathOption"
        alternatives:
          type: array
          items:
            $ref: "#/components/schemas/PathOption"
        comparison:
          type: string
          description: Why the best path may be better than the others
        source:
          type: string
          enum: [rules, ai]
      required:
        - context
        - bestPath
        - alternatives
        - comparison
        - source
    FinalVerdict:
      type: object
      properties:
        recommendedAction:
          type: string
        reasoning:
          type: string
        riskLevel:
          type: string
          enum: [low, medium, high]
      required:
        - recommendedAction
        - reasoning
        - riskLevel
...
    AstroResult:
      type: object
      properties:
        influence:
          $ref: "#/components/schemas/AstroInfluence"
        interpretation:
          type: string
        ascendant:
          type: object
          description: The rising sign at the moment of the question (the only part of the chart that is shown)
          properties:
            sign:
              type: string
            degree:
              type: number
            lord:
              type: string
          required:
            - sign
            - degree
        tuning:
          type: object
          properties:
            id:
              type: integer
              nullable: true
            astroShare:
              type: number
              nullable: true
        prashna:
          $ref: "#/components/schemas/PrashnaReading"
        timing:
          $ref: "#/components/schemas/TimingResult"
        location:
          type: object
          description: The place the chart was cast for
          properties:
            latitude:
              type: number
            longitude:
              type: number
          required:
            - latitude
            - longitude
      required:
        - influence
        - interpretation
    AnalysisResult:
      type: object
      properties:
        id:
          type: integer
        situation:
          type: string
        intent:
          $ref: "#/components/schemas/IntentResult"
        emotion:
          $ref: "#/components/schemas/EmotionResult"
        simulation:
          $ref: "#/components/schemas/SimulationResult"
        safety:
          $ref: "#/components/schemas/SafetyNotice"
        finalVerdict:
          $ref: "#/components/schemas/FinalVerdict"
        astro:
          $ref: "#/components/schemas/AstroResult"
        overallScore:
          type: number
        summary:
          type: string
        synthesis:
          $ref: "#/components/schemas/Synthesis"
        modules:
          type: array
          description: What each module did for this question (active or not, its verdict and evidence)
          items:
            $ref: "#/components/schemas/ModuleReport"
        options:
          type: object
          description: The settings this analysis was run with
          properties:
            useAi:
              type: boolean
            useAstrology:
              type: boolean
            language:
              type: string
        credits:
          $ref: "#/components/schemas/CreditsUsed"
        followUpAt:
          type: string
          description: When the app will ask how the prediction turned out
        createdAt:
          type: string
      required:
        - id
        - situation
        - intent
        - emotion
        - simulation
        - finalVerdict
        - astro
        - overallScore
        - summary
        - createdAt
...
    AnalysisRecord:
      type: object
      properties:
        id:
          type: integer
        situation:
          type: string
        intent:
          type: string
        emotion:
          type: string
        overallScore:
          type: number
        riskLevel:
          type: string
        summary:
          type: string
        fullAnalysis:
          $ref: "#/components/schemas/AnalysisResult"
        createdAt:
          type: string
      required:
        - id
        - situation
        - intent
        - emotion
        - overallScore
        - riskLevel
        - summary
        - createdAt
    AnalysisHistoryResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AnalysisRecord"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
      required:
        - items
        - total
        - limit
        - offset
...
    CostLine:
      type: object
      properties:
        key:
          type: string
          enum: [astro, ajit, manu, sivi, ai]
        label:
          type: string
        credits:
          type: integer
        note:
          type: string
      required: [key, label, credits, note]
    CostEstimate:
      type: object
      properties:
        total:
          type: integer
        aiCredits:
          type: integer
        depth:
          type: string
          enum: [standard, deep, expert]
        depthChosen:
          type: string
          enum: [auto, user]
        lines:
          type: array
          items:
            $ref: "#/components/schemas/CostLine"
        billingActive:
          type: boolean
          description: False while credits are not being charged (everything is free)
        balance:
          type: integer
        enough:
          type: boolean
      required: [total, aiCredits, depth, depthChosen, lines, billingActive, balance, enough]
    CreditsUsed:
      type: object
      description: What this analysis cost. charged is 0 while billing is off.
      properties:
        billingActive:
          type: boolean
        charged:
          type: integer
        balance:
          type: integer
          nullable: true
        depth:
          type: string
          enum: [standard, deep, expert]
        depthChosen:
          type: string
          enum: [auto, user]
        lines:
          type: array
          items:
            $ref: "#/components/schemas/CostLine"
      required: [billingActive, charged, balance, depth, depthChosen, lines]
    InsufficientCreditsResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        required:
          type: integer
        balance:
          type: integer
        cost:
          $ref: "#/components/schemas/CostEstimate"
      required: [error, message, required, balance]
...
    DeleteResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message